requirements.txt
scipy==1.11.3
chumpy==0.70
pyrender==0.1.45
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cycler==0.12.1
decorator==4.4.2
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
fsspec==2023.9.2
ftfy==6.1.1
gradio==3.47.1
gradio_client==0.6.0
h11==0.14.0
h5py==3.10.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-resources==6.1.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.1.3
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
orjson==3.9.9
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
proglog==0.1.10
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rpds-py==0.10.6
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
wcwidth==0.2.8
websockets==11.0.3
numpy==1.23.0
shapely==2.0.2
mapbox_earcut==1.0.1